💩 Basic maintenance after three years of no activity.
So... For some reason there's an error with the error message
not being a keyfield that exists. Then for some reason the
'messages' model doesn't have a keyfield for the owner of the
message and also this has been broken after the release of
Django >3.0. Eventually I will need to look into this project
again and keep it more up-to-date!